Remove duplicate accessibility service
