+++ app/src/main/AndroidManifest.xml
@@ -47,16 +47,5 @@
                 android:name="android.accessibilityservice"
                 android:resource="@xml/instagram_service_config" />
         </service>
-        <service
-            android:name=".service.ShareChooserService"
-            android:permission="android.permission.BIND_ACCESSIBILITY_SERVICE"
-            android:exported="false">
-            <intent-filter>
-                <action android:name="android.accessibilityservice.AccessibilityService" />
-            </intent-filter>
-            <meta-data
-                android:name="android.accessibilityservice"
-                android:resource="@xml/share_chooser_service_config" />
-        </service>
     </application>
 </manifest>
