+++ app/src/main/java/com/cicero/repostapp/service/InstagramPostService.kt
@@ -25,7 +25,6 @@ class InstagramPostService : AccessibilityService() {
 
     override fun onServiceConnected() {
         serviceInfo = AccessibilityServiceInfo().apply {
-            packageNames = arrayOf("com.instagram.android")
             eventTypes = AccessibilityEvent.TYPE_WINDOW_STATE_CHANGED or
                     AccessibilityEvent.TYPE_WINDOW_CONTENT_CHANGED
             feedbackType = AccessibilityServiceInfo.FEEDBACK_GENERIC
@@ -33,7 +32,17 @@ class InstagramPostService : AccessibilityService() {
     }
 
     override fun onAccessibilityEvent(event: AccessibilityEvent?) {
-        when (event?.eventType) {
+        val root = rootInActiveWindow ?: return
+
+        if (event?.packageName != "com.instagram.android") {
+            if (containsText(root, listOf("Bagikan", "Share"))) {
+                checkRememberChoice(root)
+                clickInstagramFeed(root)
+            }
+            return
+        }
+
+        when (event.eventType) {
             AccessibilityEvent.TYPE_WINDOW_STATE_CHANGED -> {
                 captionInserted = false
                 shareClicked = false
@@ -136,6 +145,42 @@ class InstagramPostService : AccessibilityService() {
             if (result != null) return result
         }
         return null
+    }
+
+    private fun findNodeByText(node: AccessibilityNodeInfo?, keywords: List<String>): AccessibilityNodeInfo? {
+        if (node == null) return null
+        for (k in keywords) {
+            if (node.text?.toString()?.contains(k, true) == true ||
+                node.contentDescription?.toString()?.contains(k, true) == true) {
+                return node
+            }
+        }
+        for (i in 0 until node.childCount) {
+            val child = node.getChild(i)
+            val res = findNodeByText(child, keywords)
+            if (res != null) return res
+        }
+        return null
+    }
+
+    private fun checkRememberChoice(root: AccessibilityNodeInfo) {
+        val node = findNodeByText(root, listOf("Ingat pilihan saya", "Remember my choice"))
+        var target = node
+        while (target != null && !target.isCheckable && !target.isClickable) {
+            target = target.parent
+        }
+        if (target != null && !target.isChecked) {
+            target.performAction(AccessibilityNodeInfo.ACTION_CLICK)
+        }
+    }
+
+    private fun clickInstagramFeed(root: AccessibilityNodeInfo) {
+        val node = findNodeByText(root, listOf("Instagram", "Feed")) ?: return
+        var target: AccessibilityNodeInfo? = node
+        while (target != null && !target.isClickable) {
+            target = target.parent
+        }
+        target?.performAction(AccessibilityNodeInfo.ACTION_CLICK)
     }
 
     override fun onInterrupt() {}
