+++ app/src/main/res/xml/instagram_service_config.xml
@@ -1,6 +1,5 @@
 <accessibility-service xmlns:android="http://schemas.android.com/apk/res/android"
     android:accessibilityEventTypes="typeWindowContentChanged|typeWindowStateChanged"
     android:accessibilityFeedbackType="feedbackGeneric"
-    android:packageNames="com.instagram.android"
     android:canRetrieveWindowContent="true"
     android:description="@string/instagram_service_description"/>
